% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMigration_stan.R
\name{fitHierarchicalMigration}
\alias{fitHierarchicalMigration}
\title{Hierarchical estimation of migration timing}
\usage{
fitHierarchicalMigration(
  data_prepped,
  initial_values = NULL,
  chains = 4,
  iter = 1000,
  ...
)
}
\description{
Hierarchical estimation of migration timing
}
