% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialValues.R
\name{getInitialValues}
\alias{getInitialValues}
\title{Get Initial values for Hierarchical Migration Estimation}
\usage{
getInitialValues(
  data,
  t_mean = 110,
  dt_mean = 20,
  mu1_mean = NULL,
  mu2_mean = NULL,
  chains = 4,
  plotme = TRUE,
  K = 1,
  ddays = 15
)
}
\description{
Get Initial values for Hierarchical Migration Estimation
}
