% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHierarchicalMigration.R
\name{plotHierarchicalMigration}
\alias{plotHierarchicalMigration}
\title{Plotting hierarchical migration fits}
\usage{
plotHierarchicalMigration(
  fit,
  pars = c("mux_mean[1]", "mux1_sigma", "muy_mean[1]", "muy1_sigma", "rho1",
    "mux_mean[2]", "mux2_sigma", "muy_mean[2]", "muy2_sigma", "rho2", "t_mean", "t_sd",
    "dt_mean", "dt_sd", "A", "sigma_ranging", "sigma_migration"),
  type = c("fit", "density", "chains", "map")
)
}
\arguments{
\item{fit}{output of \link{fitHierarchicalMigration}}

\item{plotfit}{whether to plot the fitted range ellipses and time series}

\item{plotdensity}{whether to plot the posterior densities for the chains}

\item{plotchains}{whether to plot the MCMC chains}
}
\description{
Plotting hierarchical migration fits
}
