% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{sims_fit}
\alias{sims_fit}
\title{Estimated migrations for simulated data}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(sims_fit)
}
\description{
Output of \code{\link{fitHierarchicalMigration}} function as applied to simulated migration data
\code{\link{simulated_migrations}}.
}
\details{
This is a list of three: \code{stan_data} - the data inputted into the STAN sampler, 
\code{raw_data} - the raw simple feature data, and \code{migration_fit}, the complete STAN sampler
ouput - i.e. all the MCMC chains and summary stats.
}
\keyword{data}
