% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeFit.R
\name{summarizeMigrationFit}
\alias{summarizeMigrationFit}
\title{Summarize hierarchical migration fit}
\usage{
summarizeMigrationFit(fit)
}
\arguments{
\item{fit}{output of \link{fitHierarchicalMigration}}
}
\value{
A list including estimates of migration times (start and duration) and standard deviations, 
centroid of respective starting and ending ranges, and estimated ellipses of the ranging areas before
and after the migration
}
\description{
Summarize hierarchical migration fit
}
