% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeFit.R
\name{summarizeMigrationFit}
\alias{summarizeMigrationFit}
\title{summarize fit}
\usage{
summarizeMigrationFit(fit_list)
}
\arguments{
\item{fit_list}{Output of a migration fit object}
}
\value{
A list including estimates of migration times (start and duration) and standard deviations, 
centroid of respective starting and ending ranges, and estimated ellipses of the ranging areas before
and after the migration
}
\description{
summarize fit
}
